Fixed bug in threat analysis
Applied Small fixes by mgiannopoulos24
Fixed serious bug after previous code refactoring which caused the system to make wrong
moves. In more detail, it was able to recognise the critical state of the game, but
while searching for the empty spot to place a marker, it was mixing up the rows and the
columns thus having 50% chance of playing the correct move.